Extract Value def and impls into own module
