+++ src/decoder/value.rs
@@ -1,0 +1,504 @@
+use std::rc::Rc;
+
+use num_bigint::BigInt;
+use serde::Serialize;
+
+use crate::numeric::core::{TypedConst, Value as NumValue};
+use crate::{Arith, IntRel, IntoLabel, Label, Pattern, UnaryOp};
+
+use super::{
+    MultiScope, Scope,
+    seq_kind::{SeqKind, ValueSeq},
+};
+
+impl From<NumValue> for Value {
+    fn from(value: NumValue) -> Self {
+        let NumValue::Const(c) = value;
+        Value::Numeric(Rc::new(c))
+    }
+}
+
+#[derive(Clone, PartialEq, Eq, Hash, Debug, Serialize)]
+#[serde(tag = "tag", content = "data")]
+pub enum Value {
+    Bool(bool),
+    U8(u8),
+    U16(u16),
+    U32(u32),
+    U64(u64),
+    Char(char),
+    Usize(usize),
+    // TODO[epic=embedded-num] - implement proper support for Numeric
+    Numeric(Rc<TypedConst>),
+    View {
+        offset: usize,
+    },
+    PhantomData,
+    // REVIEW - should EnumFromTo be considered a flat value?
+    EnumFromTo(std::ops::Range<usize>),
+    // vvvv Non-flat values vvvv
+    Option(Option<Box<Value>>),
+    Tuple(Vec<Value>),
+    Record(Vec<(Label, Value)>),
+    Variant(Label, Box<Value>),
+    Seq(SeqKind<Value>),
+    /// Mapped: (Raw Preimage Value, Mapped Nominal Value)
+    Mapped(Box<Value>, Box<Value>),
+    /// Branch: (Branch Index, Nominal Value)
+    Branch(usize, Box<Value>),
+    /// Poison value, used as a placeholder when a value is required but the right type cannot be induced artificially, e.g. downcasting a hard-error to a soft-error on an embedded parse
+    Poison,
+}
+
+impl From<usize> for Value {
+    fn from(value: usize) -> Value {
+        Value::Usize(value)
+    }
+}
+
+const MAX_SEQ_LEN: usize = 64;
+
+impl std::fmt::Display for Value {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            Value::Bool(b) => write!(f, "{b}"),
+            Value::U8(i) => write!(f, "{i}"),
+            Value::U16(i) => write!(f, "{i}"),
+            Value::U32(i) => write!(f, "{i}"),
+            Value::U64(i) => write!(f, "{i}"),
+            Value::Char(c) => write!(f, "{c:?}"),
+            Value::Usize(i) => write!(f, "{i}"),
+            Value::Numeric(n) => write!(f, "{n}"),
+            Value::View { offset } => write!(f, "View[+{offset}]"),
+            Value::EnumFromTo(r) => write!(f, "{r:?}"),
+            Value::Option(v) => match v {
+                None => write!(f, "None"),
+                Some(v) => write!(f, "Some({v})"),
+            },
+            Value::Tuple(vs) => {
+                write!(
+                    f,
+                    "({})",
+                    vs.iter()
+                        .map(|v| v.to_string())
+                        .collect::<Vec<_>>()
+                        .join(", ")
+                )
+            }
+            Value::Record(fields) => {
+                write!(
+                    f,
+                    "{{ {} }}",
+                    fields
+                        .iter()
+                        .map(|(f, v)| format!("{f}: {v}"))
+                        .collect::<Vec<_>>()
+                        .join(", ")
+                )
+            }
+            Value::Variant(label, value) => {
+                write!(f, "`{label}({value})")
+            }
+            Value::Seq(s_kind) => match s_kind {
+                SeqKind::Dup(n, v) => write!(f, "[{v}; {n}]"),
+                SeqKind::Strict(vs) => {
+                    if vs.len() > MAX_SEQ_LEN {
+                        write!(f, "[...; {}]", vs.len())
+                    } else {
+                        write!(
+                            f,
+                            "[{}]",
+                            vs.iter()
+                                .map(|v| v.to_string())
+                                .collect::<Vec<_>>()
+                                .join(", ")
+                        )
+                    }
+                }
+            },
+            Value::Mapped(orig, image) => {
+                write!(f, "({orig} => {image})")
+            }
+            Value::Branch(n, value) => write!(f, "({n} :~ {value})"),
+            // REVIEW - is this over-verbose?
+            Value::PhantomData => write!(f, "<PhantomData>"),
+            Value::Poison => write!(f, "NO_VALUE"),
+        }
+    }
+}
+
+impl Value {
+    pub(crate) fn tuple_proj(&self, index: usize) -> &Self {
+        match self.coerce_mapped_value() {
+            Value::Tuple(vs) => &vs[index],
+            Value::Poison => self,
+            _ => panic!("expected tuple"),
+        }
+    }
+
+    pub(crate) fn record_proj(&self, label: &str) -> &Self {
+        match self {
+            Value::Poison => self,
+            Value::Record(fields) => match fields.iter().find(|(l, _)| label == l) {
+                Some((_, v)) => v,
+                None => panic!("{label} not found in record"),
+            },
+            other => panic!("expected record, found {other:?}"),
+        }
+    }
+
+    // TODO[epic=embedded-num] - no support for negative values or signed constants yet
+    pub(crate) fn matches_inner<'a>(
+        &'a self,
+        scope: &mut MultiScope<'a>,
+        pattern: &Pattern,
+    ) -> bool {
+        match pattern {
+            Pattern::Binding(name) => {
+                scope.push(name.clone(), self);
+                true
+            }
+            Pattern::Wildcard => true,
+            Pattern::Bool(b0) => matches!(self, Value::Bool(b1) if b0 == b1),
+            Pattern::U8(i0) => match self {
+                Value::U8(i1) if i0 == i1 => true,
+                Value::Numeric(n) => n.matches_u8(*i0),
+                _ => false,
+            },
+            Pattern::U16(i0) => match self {
+                Value::U16(i1) if i0 == i1 => true,
+                Value::Numeric(n) => n.matches_u16(*i0),
+                _ => false,
+            },
+            Pattern::U32(i0) => match self {
+                Value::U32(i1) if i0 == i1 => true,
+                Value::Numeric(n) => n.matches_u32(*i0),
+                _ => false,
+            },
+            Pattern::U64(i0) => match self {
+                Value::U64(i1) if i0 == i1 => true,
+                Value::Numeric(n) => n.matches_u64(*i0),
+                _ => false,
+            },
+            Pattern::Int(bounds) => match self {
+                Value::U8(n) => bounds.contains(usize::from(*n)),
+                Value::U16(n) => bounds.contains(usize::from(*n)),
+                Value::U32(n) => bounds.contains(usize::try_from(*n).unwrap()),
+                Value::U64(n) => bounds.contains(usize::try_from(*n).unwrap()),
+                Value::Usize(n) => bounds.contains(*n),
+                Value::Numeric(n) => n.matches_int_range(*bounds),
+                _ => false,
+            },
+            Pattern::ZConst(z) => match self {
+                Value::U8(n) => z == &BigInt::from(*n),
+                Value::U16(n) => z == &BigInt::from(*n),
+                Value::U32(n) => z == &BigInt::from(*n),
+                Value::U64(n) => z == &BigInt::from(*n),
+                Value::Usize(n) => z == &BigInt::from(*n),
+                Value::Numeric(n) => n.eq_num(z),
+                _ => false,
+            },
+            Pattern::ZRange(range) => match self {
+                Value::U8(n) => range.contains(&BigInt::from(*n)),
+                Value::U16(n) => range.contains(&BigInt::from(*n)),
+                Value::U32(n) => range.contains(&BigInt::from(*n)),
+                Value::U64(n) => range.contains(&BigInt::from(*n)),
+                Value::Usize(n) => range.contains(&BigInt::from(*n)),
+                Value::Numeric(n) => range.contains(n.as_raw_value()),
+                _ => false,
+            },
+            Pattern::Char(c0) => matches!(self, Value::Char(c1) if c0 == c1),
+            Pattern::Tuple(ps) => match self {
+                Value::Tuple(vs) if ps.len() == vs.len() => {
+                    for (p, v) in Iterator::zip(ps.iter(), vs.iter()) {
+                        if !v.matches_inner(scope, p) {
+                            return false;
+                        }
+                    }
+                    true
+                }
+                _ => false,
+            },
+            Pattern::Seq(ps) => match self {
+                Value::Seq(vs) if ps.len() == vs.len() => {
+                    for (p, v) in Iterator::zip(ps.iter(), vs.iter()) {
+                        if !v.matches_inner(scope, p) {
+                            return false;
+                        }
+                    }
+                    true
+                }
+                _ => false,
+            },
+
+            Pattern::Variant(label0, p) => match self {
+                Value::Variant(label1, v) if label0 == label1 => v.matches_inner(scope, p),
+                _ => false,
+            },
+            Pattern::Option(None) => matches!(self, Value::Option(None)),
+            Pattern::Option(Some(p)) => match self {
+                Value::Option(Some(v)) => v.matches_inner(scope, p),
+                _ => false,
+            },
+        }
+    }
+
+    pub(crate) fn matches<'a>(
+        &'a self,
+        scope: &'a Scope<'a>,
+        pattern: &Pattern,
+    ) -> Option<MultiScope<'a>> {
+        let mut pattern_scope = MultiScope::new(scope);
+        self.coerce_mapped_value()
+            .matches_inner(&mut pattern_scope, pattern)
+            .then_some(pattern_scope)
+    }
+
+    pub fn coerce_mapped_value(&self) -> &Self {
+        match self {
+            Value::Mapped(_orig, v) => v.coerce_mapped_value(),
+            Value::Branch(_n, v) => v.coerce_mapped_value(),
+            v => v,
+        }
+    }
+
+    pub fn extract_mapped_value(self) -> Self {
+        match self {
+            Value::Mapped(_orig, v) => v.extract_mapped_value(),
+            Value::Branch(_n, v) => v.extract_mapped_value(),
+            v => v,
+        }
+    }
+
+    pub(crate) fn get_sequence(&self) -> Option<ValueSeq<'_, Self>> {
+        match self {
+            Value::Seq(elts) => Some(ValueSeq::ValueSeq(elts)),
+            Value::EnumFromTo(range) => Some(ValueSeq::IntRange(range.clone())),
+            _ => None,
+        }
+    }
+
+    pub(crate) fn is_boolean(&self) -> bool {
+        matches!(self.coerce_mapped_value(), Value::Bool(_))
+    }
+}
+
+impl Value {
+    pub const UNIT: Value = Value::Tuple(Vec::new());
+
+    pub fn record<Name: IntoLabel>(fields: impl IntoIterator<Item = (Name, Value)>) -> Value {
+        Value::Record(
+            fields
+                .into_iter()
+                .map(|(label, value)| (label.into(), value))
+                .collect(),
+        )
+    }
+
+    pub fn variant(label: impl IntoLabel, value: impl Into<Box<Value>>) -> Value {
+        Value::Variant(label.into(), value.into())
+    }
+
+    /// Unwraps any compatible numeric-typed `Value` and returns the contained number as a `usize`.
+    ///
+    /// # Panics
+    ///
+    /// Panics if the value is not numeric.
+    ///
+    /// May additionally panic in rare cases such as `u64`-to-`usize` conversion on 32-bit architectures.
+    pub(crate) fn unwrap_usize(&self) -> usize {
+        match self {
+            Value::U8(n) => usize::from(*n),
+            Value::U16(n) => usize::from(*n),
+            Value::U32(n) => usize::try_from(*n).unwrap(),
+            Value::U64(n) => usize::try_from(*n).unwrap(),
+            Value::Usize(n) => *n,
+            Value::Numeric(_tc) => {
+                unimplemented!("unwrap_usize: conversion from TypedConst not yet supported")
+            }
+            other => panic!("value is not a number: {other:?}"),
+        }
+    }
+
+    /// Unwraps `Value::U8` and returns the contained value, or panics if the value is not `Value::U8`.
+    pub(crate) fn get_as_u8(&self) -> u8 {
+        match self {
+            Value::U8(n) => *n,
+            Value::U16(..) | Value::U32(..) | Value::U64(..) | Value::Usize(..) => panic!(
+                "value is numeric but not u8 (this may be a soft error, or even success, in future)"
+            ),
+            Value::Numeric(_tc) => {
+                unimplemented!("get_as_u8: conversion from TypedConst not yet supported")
+            }
+            _ => panic!("value is not a number"),
+        }
+    }
+
+    pub(crate) fn unwrap_tuple(self) -> Vec<Value> {
+        match self {
+            Value::Tuple(values) => values,
+            _ => panic!("value is not a tuple"),
+        }
+    }
+
+    pub(crate) fn unwrap_bool(&self) -> bool {
+        match self {
+            Value::Bool(b) => *b,
+            _ => panic!("value is not a bool"),
+        }
+    }
+
+    /// FIXME - do we really need this?
+    #[allow(dead_code)]
+    fn unwrap_char(self) -> char {
+        match self {
+            Value::Char(c) => c,
+            _ => panic!("value is not a char"),
+        }
+    }
+}
+
+fn __rel<T>(rel: IntRel, left: T, right: T) -> bool
+where
+    T: Eq + Ord,
+{
+    match rel {
+        IntRel::Eq => left == right,
+        IntRel::Ne => left != right,
+        IntRel::Gt => left > right,
+        IntRel::Gte => left >= right,
+        IntRel::Lt => left < right,
+        IntRel::Lte => left <= right,
+    }
+}
+
+fn __arith<T>(arith: Arith, left: T, right: T) -> T
+where
+    T: num_traits::CheckedAdd,
+    T: num_traits::CheckedSub,
+    T: num_traits::CheckedMul,
+    T: num_traits::CheckedDiv,
+    T: num_traits::CheckedRem,
+    T: num_traits::CheckedShl,
+    T: num_traits::CheckedShr,
+    T: num_traits::AsPrimitive<u32>,
+    T: std::ops::BitOr<Output = T>,
+    T: std::ops::BitAnd<Output = T>,
+{
+    match arith {
+        Arith::Add => left
+            .checked_add(&right)
+            .unwrap_or_else(|| panic!("integer overflow")),
+        Arith::Sub => left
+            .checked_sub(&right)
+            .unwrap_or_else(|| panic!("integer overflow")),
+        Arith::Mul => left
+            .checked_mul(&right)
+            .unwrap_or_else(|| panic!("integer overflow")),
+        Arith::Div => left
+            .checked_div(&right)
+            .unwrap_or_else(|| panic!("integer overflow")),
+        Arith::Rem => left
+            .checked_rem(&right)
+            .unwrap_or_else(|| panic!("integer overflow")),
+        Arith::Shl => left << right.as_(),
+        Arith::Shr => left >> right.as_(),
+        Arith::BitOr => left | right,
+        Arith::BitAnd => left & right,
+        Arith::BoolOr | Arith::BoolAnd => unreachable!("bool ops should be handled separately"),
+    }
+}
+
+fn __unary<T>(op: UnaryOp, value: T) -> T
+where
+    T: num_traits::CheckedAdd,
+    T: num_traits::CheckedSub,
+    T: num_traits::One,
+{
+    match op {
+        UnaryOp::IntPred => value.checked_sub(&T::one()).unwrap_or_else(|| {
+            panic!(
+                "__unary::<{}>@IntPred: integer underflow",
+                std::any::type_name::<T>()
+            )
+        }),
+        UnaryOp::IntSucc => value.checked_add(&T::one()).unwrap_or_else(|| {
+            panic!(
+                "__unary::<{}>@IntSucc: integer overflow",
+                std::any::type_name::<T>()
+            )
+        }),
+        UnaryOp::BoolNot => unreachable!("bool ops should be handled separately"),
+    }
+}
+
+impl Value {
+    pub fn int_rel(rel: IntRel, left: Value, right: Value) -> Value {
+        match (left, right) {
+            (Value::U8(l), Value::U8(r)) => Value::Bool(__rel(rel, l, r)),
+            (Value::U16(l), Value::U16(r)) => Value::Bool(__rel(rel, l, r)),
+            (Value::U32(l), Value::U32(r)) => Value::Bool(__rel(rel, l, r)),
+            (Value::U64(l), Value::U64(r)) => Value::Bool(__rel(rel, l, r)),
+            (Value::Usize(l), Value::Usize(r)) => Value::Bool(__rel(rel, l, r)),
+            (Value::Numeric(_l), Value::Numeric(_r)) => {
+                unimplemented!("int_rel: TypedConst comparison not yet implemented");
+            }
+            (left, right) => {
+                panic!("cannot apply int-rel {rel:?} to (`{left:?}`, `{right:?}`)")
+            }
+        }
+    }
+
+    pub fn arith(arith: Arith, left: Value, right: Value) -> Value {
+        if matches!(arith, Arith::BoolOr | Arith::BoolAnd) {
+            match (left, right) {
+                (Value::Bool(l), Value::Bool(r)) => match arith {
+                    Arith::BoolOr => Value::Bool(l || r),
+                    Arith::BoolAnd => Value::Bool(l && r),
+                    _ => unreachable!(),
+                },
+                (left, right) => {
+                    panic!(
+                        "cannot perform boolean operation on non-boolean operand-pair (`{left:?}`, `{right:?}`)"
+                    );
+                }
+            }
+        } else {
+            match (left, right) {
+                (Value::U8(l), Value::U8(r)) => Value::U8(__arith(arith, l, r)),
+                (Value::U16(l), Value::U16(r)) => Value::U16(__arith(arith, l, r)),
+                (Value::U32(l), Value::U32(r)) => Value::U32(__arith(arith, l, r)),
+                (Value::U64(l), Value::U64(r)) => Value::U64(__arith(arith, l, r)),
+                (Value::Usize(l), Value::Usize(r)) => Value::Usize(__arith(arith, l, r)),
+                (Value::Numeric(_l), Value::Numeric(_r)) => {
+                    panic!(
+                        "raw arithmetic on numerics should be done in numeric model, or with Expr-level casts beforehand"
+                    );
+                }
+                (left, right) => {
+                    panic!("cannot apply arith {arith:?} to (`{left:?}`, `{right:?}`)")
+                }
+            }
+        }
+    }
+
+    pub fn unary(op: UnaryOp, value: Value) -> Value {
+        match op {
+            UnaryOp::BoolNot => match value {
+                Value::Bool(b) => Value::Bool(!b),
+                _ => panic!("cannot apply bool-not to non-boolean operand (`{value:?}`)"),
+            },
+            op => match value {
+                Value::U8(i) => Value::U8(__unary(op, i)),
+                Value::U16(i) => Value::U16(__unary(op, i)),
+                Value::U32(i) => Value::U32(__unary(op, i)),
+                Value::U64(i) => Value::U64(__unary(op, i)),
+                Value::Usize(i) => Value::Usize(__unary(op, i)),
+                Value::Numeric(_i) => {
+                    panic!("top-level unary operations should not be performed on raw-numeric");
+                }
+                _ => panic!("cannot apply unary {op:?} to non-numeric operand (`{value:?}`)"),
+            },
+        }
+    }
+}
